Move send message test
