+++ python/jupyterlab-collaborative-chat/ui-tests/tests/jupyterlab_collaborative_chat.spec.ts
@@ -11,9 +11,14 @@ import {
 } from '@jupyterlab/galata';
 import { Contents, User } from '@jupyterlab/services';
 import { ReadonlyJSONObject, UUID } from '@lumino/coreutils';
-import { Locator } from '@playwright/test';
 
-import { openChat, openChatToSide, sendMessage, splitMainArea, USER } from './test-utils';
+import {
+  openChat,
+  openChatToSide,
+  openSettings,
+  sendMessage,
+  USER
+} from './test-utils';
 
 const FILENAME = 'my-chat.chat';
 const MSG_CONTENT = 'Hello World!';
@@ -41,26 +46,6 @@ const readFileContent = async (
   return await page.evaluate(async filepath => {
     return await window.jupyterapp.serviceManager.contents.get(filepath);
   }, filename);
-};
-
-const openSettings = async (
-  page: IJupyterLabPageFixture,
-  globalSettings?: boolean
-): Promise<Locator> => {
-  const args = globalSettings ? {} : { query: 'Chat' };
-  await page.evaluate(async args => {
-    await window.jupyterapp.commands.execute('settingeditor:open', args);
-  }, args);
-
-  // Activate the settings tab, sometimes it does not automatically.
-  const settingsTab = page
-    .getByRole('main')
-    .getByRole('tab', { name: 'Settings', exact: true });
-  await settingsTab.click();
-  await page.waitForCondition(
-    async () => (await settingsTab.getAttribute('aria-selected')) === 'true'
-  );
-  return (await page.activity.getPanelLocator('Settings')) as Locator;
 };
 
 test.describe('#commandPalette', () => {
@@ -212,245 +197,6 @@ test.describe('#settings', () => {
       name: 'stackMessages'
     });
     expect(stackMessages!).toBeChecked();
-  });
-});
-
-test.describe('#sendMessages', () => {
-  test.beforeEach(async ({ page }) => {
-    // Create a chat file
-    await page.filebrowser.contents.uploadContent('{}', 'text', FILENAME);
-  });
-
-  test.afterEach(async ({ page }) => {
-    if (await page.filebrowser.contents.fileExists(FILENAME)) {
-      await page.filebrowser.contents.deleteFile(FILENAME);
-    }
-  });
-
-  test('should send a message using button', async ({ page }) => {
-    const chatPanel = await openChat(page, FILENAME);
-    const input = chatPanel
-      .locator('.jp-chat-input-container')
-      .getByRole('combobox');
-    const sendButton = chatPanel.locator(
-      '.jp-chat-input-container .jp-chat-send-button'
-    );
-    await input.pressSequentially(MSG_CONTENT);
-    await sendButton.click();
-
-    const messages = chatPanel.locator('.jp-chat-messages-container');
-    await expect(messages.locator('.jp-chat-message')).toHaveCount(1);
-    // It seems that the markdown renderer adds a new line.
-    await expect(
-      messages.locator('.jp-chat-message .jp-chat-rendermime-markdown')
-    ).toHaveText(MSG_CONTENT + '\n');
-  });
-
-  test('should send a message using keyboard', async ({ page }) => {
-    const chatPanel = await openChat(page, FILENAME);
-    const input = chatPanel
-      .locator('.jp-chat-input-container')
-      .getByRole('combobox');
-    await input.pressSequentially(MSG_CONTENT);
-    await input.press('Enter');
-
-    const messages = chatPanel.locator('.jp-chat-messages-container');
-    await expect(messages.locator('.jp-chat-message')).toHaveCount(1);
-    // It seems that the markdown renderer adds a new line.
-    await expect(
-      messages.locator('.jp-chat-message .jp-chat-rendermime-markdown')
-    ).toHaveText(MSG_CONTENT + '\n');
-  });
-
-  test('should use settings value sendWithShiftEnter', async ({ page }) => {
-    // Modify the settings
-    const settings = await openSettings(page);
-    const sendWithShiftEnter = settings?.getByRole('checkbox', {
-      name: 'sendWithShiftEnter'
-    });
-    await sendWithShiftEnter?.check();
-
-    // wait for the settings to be saved
-    await expect(page.activity.getTabLocator('Settings')).toHaveAttribute(
-      'class',
-      /jp-mod-dirty/
-    );
-    await expect(page.activity.getTabLocator('Settings')).not.toHaveAttribute(
-      'class',
-      /jp-mod-dirty/
-    );
-
-    // Should not send message with Enter
-    const chatPanel = await openChat(page, FILENAME);
-    const messages = chatPanel.locator('.jp-chat-messages-container');
-    const input = chatPanel
-      .locator('.jp-chat-input-container')
-      .getByRole('combobox');
-    await input!.pressSequentially(MSG_CONTENT);
-    await input!.press('Enter');
-
-    await expect(messages!.locator('.jp-chat-message')).toHaveCount(0);
-
-    // Should not send message with Shift+Enter
-    await input!.press('Shift+Enter');
-    await expect(messages!.locator('.jp-chat-message')).toHaveCount(1);
-
-    // It seems that the markdown renderer adds a new line, but the '\n' inserter when
-    // pressing Enter above is trimmed.
-    await expect(
-      messages.locator('.jp-chat-message .jp-chat-rendermime-markdown')
-    ).toHaveText(MSG_CONTENT + '\n');
-  });
-
-  test('should update settings value sendWithShiftEnter on existing chat', async ({
-    page
-  }) => {
-    const chatPanel = await openChat(page, FILENAME);
-
-    // Modify the settings
-    const settings = await openSettings(page);
-    const sendWithShiftEnter = settings?.getByRole('checkbox', {
-      name: 'sendWithShiftEnter'
-    });
-    await sendWithShiftEnter?.check();
-
-    // wait for the settings to be saved
-    await expect(page.activity.getTabLocator('Settings')).toHaveAttribute(
-      'class',
-      /jp-mod-dirty/
-    );
-    await expect(page.activity.getTabLocator('Settings')).not.toHaveAttribute(
-      'class',
-      /jp-mod-dirty/
-    );
-
-    // Activate the chat panel
-    await page.activity.activateTab(FILENAME);
-
-    // Should not send message with Enter
-    const messages = chatPanel.locator('.jp-chat-messages-container');
-    const input = chatPanel
-      .locator('.jp-chat-input-container')
-      .getByRole('combobox');
-    await input!.pressSequentially(MSG_CONTENT);
-    await input!.press('Enter');
-
-    await expect(messages!.locator('.jp-chat-message')).toHaveCount(0);
-
-    // Should not send message with Shift+Enter
-    await input!.press('Shift+Enter');
-    await expect(messages!.locator('.jp-chat-message')).toHaveCount(1);
-
-    // It seems that the markdown renderer adds a new line, but the '\n' inserted when
-    // pressing Enter above is trimmed.
-    await expect(
-      messages.locator('.jp-chat-message .jp-chat-rendermime-markdown')
-    ).toHaveText(MSG_CONTENT + '\n');
-  });
-
-  test('should disable send with selection when there is no notebook', async ({
-    page
-  }) => {
-    const chatPanel = await openChat(page, FILENAME);
-    const input = chatPanel
-      .locator('.jp-chat-input-container')
-      .getByRole('combobox');
-    const openerButton = chatPanel.locator(
-      '.jp-chat-input-container .jp-chat-send-include-opener'
-    );
-    const sendWithSelection = page.locator('.jp-chat-send-include');
-
-    await input.pressSequentially(MSG_CONTENT);
-    await openerButton.click();
-    await expect(sendWithSelection).toBeVisible();
-    await expect(sendWithSelection).toBeDisabled();
-    await expect(sendWithSelection).toContainText(
-      'No selection or active cell'
-    );
-  });
-
-  test('should send with cell content', async ({ page }) => {
-    const cellContent = 'a = 1\nprint(f"a={a}")';
-    const chatPanel = await openChat(page, FILENAME);
-    const messages = chatPanel.locator('.jp-chat-messages-container');
-    const input = chatPanel
-      .locator('.jp-chat-input-container')
-      .getByRole('combobox');
-    const openerButton = chatPanel.locator(
-      '.jp-chat-input-container .jp-chat-send-include-opener'
-    );
-    const sendWithSelection = page.locator('.jp-chat-send-include');
-
-    const notebook = await page.notebook.createNew();
-    // write content in the first cell.
-    const cell = (await page.notebook.getCellLocator(0))!;
-    await cell.getByRole('textbox').pressSequentially(cellContent);
-
-    await splitMainArea(page, notebook!);
-
-    await input.pressSequentially(MSG_CONTENT);
-    await openerButton.click();
-    await expect(sendWithSelection).toBeVisible();
-    await expect(sendWithSelection).toBeEnabled();
-    await expect(sendWithSelection).toContainText('Code from 1 active cell');
-    await sendWithSelection.click();
-
-    await expect(messages!.locator('.jp-chat-message')).toHaveCount(1);
-
-    // It seems that the markdown renderer adds a new line, but the '\n' inserter when
-    // pressing Enter above is trimmed.
-    await expect(
-      messages.locator('.jp-chat-message .jp-chat-rendermime-markdown')
-    ).toHaveText(`${MSG_CONTENT}\n${cellContent}\n`);
-  });
-
-  test('should send with text selection', async ({ page }) => {
-    const cellContent = 'a = 1\nprint(f"a={a}")';
-    const chatPanel = await openChat(page, FILENAME);
-    const messages = chatPanel.locator('.jp-chat-messages-container');
-    const input = chatPanel
-      .locator('.jp-chat-input-container')
-      .getByRole('combobox');
-    const openerButton = chatPanel.locator(
-      '.jp-chat-input-container .jp-chat-send-include-opener'
-    );
-    const sendWithSelection = page.locator('.jp-chat-send-include');
-
-    const notebook = await page.notebook.createNew();
-    await splitMainArea(page, notebook!);
-
-    // write content in the first cell.
-    const cell = (await page.notebook.getCellLocator(0))!;
-    await cell.getByRole('textbox').pressSequentially(cellContent);
-
-    // wait for code mirror to be ready.
-    await expect(cell.locator('.cm-line')).toHaveCount(2);
-    await expect(
-      cell.locator('.cm-line').nth(1).locator('.cm-builtin')
-    ).toBeAttached();
-
-    // select the 'print' statement in the second line.
-    const selection = cell
-      ?.locator('.cm-line')
-      .nth(1)
-      .locator('.cm-builtin')
-      .first();
-    await selection.dblclick({ position: { x: 10, y: 10 } });
-
-    await input.pressSequentially(MSG_CONTENT);
-    await openerButton.click();
-    await expect(sendWithSelection).toBeVisible();
-    await expect(sendWithSelection).toBeEnabled();
-    await expect(sendWithSelection).toContainText('1 line(s) selected');
-    await sendWithSelection.click();
-
-    await expect(messages!.locator('.jp-chat-message')).toHaveCount(1);
-
-    // It seems that the markdown renderer adds a new line, but the '\n' inserter when
-    // pressing Enter above is trimmed.
-    await expect(
-      messages.locator('.jp-chat-message .jp-chat-rendermime-markdown')
-    ).toHaveText(`${MSG_CONTENT}\nprint\n`);
   });
 });
