+++ python/jupyterlab-collaborative-chat/ui-tests/tests/test-utils.ts
@@ -87,3 +87,23 @@ export const splitMainArea = async (
   await page.mouse.move(viewerBBox!.x + 0.5 * viewerBBox!.width, 600);
   await page.mouse.up();
 };
+
+export const openSettings = async (
+  page: IJupyterLabPageFixture,
+  globalSettings?: boolean
+): Promise<Locator> => {
+  const args = globalSettings ? {} : { query: 'Chat' };
+  await page.evaluate(async args => {
+    await window.jupyterapp.commands.execute('settingeditor:open', args);
+  }, args);
+
+  // Activate the settings tab, sometimes it does not automatically.
+  const settingsTab = page
+    .getByRole('main')
+    .getByRole('tab', { name: 'Settings', exact: true });
+  await settingsTab.click();
+  await page.waitForCondition(
+    async () => (await settingsTab.getAttribute('aria-selected')) === 'true'
+  );
+  return (await page.activity.getPanelLocator('Settings')) as Locator;
+};
